@@ -41,11 +41,12 @@ dependencies:
   image_picker: ^0.8.6+1
   path_provider: ^2.0.12
   font_awesome_flutter: ^10.4.0
-  cached_network_image: ^3.2.3
+  cached_network_image: ^3.4.1
   url_launcher: ^6.1.12
   flutter_dotenv: ^5.1.0
-  sentry_flutter: ^7.10.1
+  sentry_flutter: ^8.7.0
   package_info_plus: ^4.2.0
+  google_sign_in: ^6.2.1
 
 dev_dependencies:
   flutter_test:
