@@ -12,8 +12,8 @@ import 'package:chicpic/models/auth/login_user_data.dart';
 
 import 'package:chicpic/ui/auth/widgets/submit_button.dart';
 
-// import 'package:chicpic/ui/auth/widgets/auth_divider.dart';
-// import 'package:chicpic/ui/auth/widgets/google_button.dart';
+import 'package:chicpic/ui/auth/widgets/auth_divider.dart';
+import 'package:chicpic/ui/auth/widgets/google_button.dart';
 
 class LoginScreen extends StatefulWidget {
   const LoginScreen({Key? key}) : super(key: key);
@@ -148,11 +148,8 @@ class _LoginScreenState extends State<LoginScreen> {
                                   )
                                 : null,
                           ),
-                          // const AuthDivider(),
-                          // GoogleButton(
-                          //   onPressed: state is LoginLoading ? null : () {},
-                          //   label: 'Sign in with Google',
-                          // ),
+                          const AuthDivider(),
+                          const GoogleButton(),
                         ],
                       );
                     },
